+++ ginga/util/viewer.py
@@ -64,6 +64,10 @@ class ViewerBase(Callback.Callbacks):
     def __str__(self):
         return self.name
 
+    def set_name(self, name):
+        """Set viewer name."""
+        self.name = name
+
     def get_settings(self):
         """Get the settings used by this instance.